requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
python-dotenv==1.0.0
firebase-admin==6.2.0
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0
google-generativeai==0.8.3
requests==2.31.0
Pillow==10.0.1
gunicorn==21.2.0

# Security
cryptography==41.0.7
bcrypt==4.1.2
PyJWT==2.8.0
flask-cors==4.0.0
flask-limiter==3.5.0

# AI Services
openai==1.3.0
replicate==0.22.0

# Testing
pytest==7.4.2
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-flask==1.3.0

# Code Quality
black==23.7.0
flake8==6.0.0
mypy==1.7.0
isort==5.12.0

# Monitoring
sentry-sdk[flask]==1.38.0
prometheus-client==0.19.0

# Caching
redis==5.0.1
flask-caching==2.1.0
